Most of  these can already be found here
Ill try  to get to putting up the DDL links like the other one, but it's just damn it gets old and I aint  that workflow smart
The bat's are nothing special nic just calls up network interfaces direct without looking and clicking at shit. ENV in this set or not is same but pulls up the GUI for system vars. PurgeDRIVEFS is meant, in a sense to be ran along with teh registry part...The DriveFS registry limits google to fuck all basically for cache. The DriveFS Batch could be standalone too, if you want cache, but if you have A LOT of frequent changes, drivefs will hang on winders, so I just run that and it purges all directories etc.  (after it's closed/toggle dont autorun or itll fail) Don't think that's the one that'll autolaunch too but 1 click aint bad. DO NOT RUN THAT BAT IF YOU DO USE CACHING AS IT GOES BYEBYE viA RCLONE PURGE. If you're safe and good  to go by all means have at it. That's why i was hinting at also using that registry hack. Might be able to find sauce for it but google yanked ze post.